requirements.txt
Django>=4.0.4,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.22.1,<0.23
Pillow>=9.1.0,<9.2
uwsgi>=2.0.20,<2.1
djangorestframework-simplejwt==5.2.2
django-auditlog==2.2.2
celery==5.2.7
redis==4.5.4
django-celery-beat==2.5.0
